Add files via upload
One of the key features of the Sorting Visualizer is its customizability. Users have the flexibility to adjust the size of the input array, allowing them to visualize the sorting process for various data set sizes. Additionally, users can control the speed of the visualization, choosing to slow it down or speed it up. This feature enables users to closely analyze and comprehend the intricacies of each algorithm's execution.

The Sorting Visualizer offers interactive controls that empower users to pause, resume, or reset the sorting process at any point. This functionality allows users to closely examine the changes made at each step and better grasp the inner workings of the sorting algorithm. Furthermore, each element in the array is represented by a visually appealing bar with varying heights and colors, providing a colorful representation that aids in the understanding of the sorting process.

To enhance the educational value of the project, the Sorting Visualizer includes concise descriptions of each sorting algorithm. These descriptions outline the working principles, time complexities, and notable characteristics of each algorithm. This information equips users with a deeper understanding of the underlying logic and efficiency of the algorithms.

Using the Sorting Visualizer is straightforward. Users can clone the repository or download the source code and open it in a web browser. From there, they can select their desired sorting algorithm, customize the array size and visualization speed, and start the visualization. The interactive controls allow users to closely observe the sorting process and interact as needed.

Overall, the Sorting Visualizer is an educational and insightful project that provides a visually engaging and interactive platform for learning and comprehending sorting algorithms. Its features, customizability, and intuitive interface make it an invaluable resource for students, developers, and algorithm enthusiasts seeking to deepen their understanding of sorting algorithms and their implementations.